LayoutTests/webanimations/document-timeline.html

<script src="../resources/js-test-pre.js"></script>
<script>

description("Check that the DocumentTimeline interface is defined and that the Document has one by default.");

debug("The DocumentTimeline interface is defined.");
shouldBeDefined("DocumentTimeline");

debug("");
debug("The Document interface exposes a timeline property.");
shouldBeDefined(document.timeline);

debug("");
debug("The document timeline is a DocumentTimeline.");
shouldBeType("document.timeline", "DocumentTimeline");

debug("");
debug("The document timeline inherits from AnimationTimeline.");
shouldBeType("document.timeline", "AnimationTimeline");

debug("");
debug("The document timeline is a single object.");
const timeline = document.timeline;
shouldBe("timeline", "document.timeline");

debug("");

</script>
<script src="../resources/js-test-post.js"></script>
