LayoutTests/webanimations/document-timeline-expected.txt

Check that the DocumentTimeline interface is defined and that the Document has one by default.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


The DocumentTimeline interface is defined.
PASS DocumentTimeline is defined.

The Document interface exposes a timeline property.
PASS [object DocumentTimeline] is defined.

The document timeline is a DocumentTimeline.
PASS document.timeline is an instance of DocumentTimeline

The document timeline inherits from AnimationTimeline.
PASS document.timeline is an instance of AnimationTimeline

The document timeline is a single object.
PASS timeline is document.timeline

PASS successfullyParsed is true

TEST COMPLETE
